test: sort test files in alphanum order
The order returned by `find` is not guaranteed to be in any
order, and may change between PRs.  This makes it difficult to reproduce
CI test runs.  If some failures are test order dependent, then
you may see failures in a PR that you did not see in previous PRs.
A stable test order is needed.  The `sort` command will sort in
alphanum (ASCII) order.

Signed-off-by: Rich Megginson <rmeggins@redhat.com>